+++ test/conversion.test.js
@@ -2,11 +2,15 @@ var assert = require('assert');
 var typed = require('../typed-function');
 var strictEqualArray = require('./strictEqualArray');
 
+function convertBool (b) {
+  return +b;
+}
+
 describe('conversion', function () {
 
   before(function () {
-    typed.conversions = [
-      {from: 'boolean', to: 'number', convert: function (x) {return +x;}},
+    typed.addConversions([
+      {from: 'boolean', to: 'number', convert: convertBool},
       {from: 'boolean', to: 'string', convert: function (x) {return x + '';}},
       {from: 'number',  to: 'string', convert: function (x) {return x + '';}},
       {
@@ -18,12 +22,12 @@ describe('conversion', function () {
         },
         fallible: true // TODO: not yet supported
       }
-    ];
+    ]);
   });
 
   after(function () {
     // cleanup conversions
-    typed.conversions = [];
+    typed.clearConversions();
   });
 
   it('should add conversions to a function with one argument', function() {
@@ -49,12 +53,14 @@ describe('conversion', function () {
       }
     };
 
-    assert.equal(typed2.conversions.length, 0);
+    assert.strictEqual(typed2._findType('string').conversionsTo.length, 0);
 
     typed2.addConversion(conversion);
 
-    assert.equal(typed2.conversions.length, 1);
-    assert.strictEqual(typed2.conversions[0], conversion);
+    assert.strictEqual(typed2._findType('string').conversionsTo.length, 1);
+    assert.strictEqual(
+      typed2._findType('string').conversionsTo[0].convert,
+      conversion.convert);
   });
 
   it('should throw an error when passing an invalid conversion object to addConversion', function() {
@@ -68,6 +74,14 @@ describe('conversion', function () {
     assert.throws(function () {typed2.addConversion({from: 2, to: 'string', convert: function () {}})}, errMsg);
     assert.throws(function () {typed2.addConversion({from: 'number', to: 2, convert: function () {}})}, errMsg);
     assert.throws(function () {typed2.addConversion({from: 'number', to: 'string', convert: 'foo'})}, errMsg);
+  });
+
+  it('should throw an error when attempting to add a conversion to unknown type', function () {
+    assert.throws(() => typed.addConversion({
+      from: 'number',
+      to: 'garbage',
+      convert: () => null
+    }), /Unknown type/);
   });
 
   it('should add conversions to a function with multiple arguments', function() {
@@ -95,10 +109,9 @@ describe('conversion', function () {
     assert.equal(fn('foo', true), 'string, number');
     assert.equal(fn('foo', 2), 'string, number');
     assert.equal(fn('foo', 'foo'), 'string, string');
-    assert.deepEqual(Object.keys(fn.signatures), [
-      'string,number',
-      'string,string'
-    ]);
+    assert.equal(Object.keys(fn.signatures).length, 2);
+    assert.ok('string,number' in fn.signatures);
+    assert.ok('string,string' in fn.signatures);
   });
 
   it('should add conversions to a function with rest parameters (1)', function() {
@@ -160,11 +173,11 @@ describe('conversion', function () {
 
   it('should add conversions to a function with rest parameters in a non-conflicting way', function() {
     var typed2 = typed.create();
-    typed2.conversions = [
+    typed2.addConversions([
       {from: 'boolean', to: 'number', convert: function (x) {return +x}},
       {from: 'string',  to: 'number', convert: function (x) {return parseFloat(x)}},
       {from: 'string', to: 'boolean', convert: function (x) {return !!x}}
-    ];
+    ]);
 
     // booleans can be converted to numbers, so the `...number` signature
     // will match. But the `...boolean` signature is a better (exact) match so that
@@ -202,9 +215,9 @@ describe('conversion', function () {
 
   it('should order conversions and type Object correctly ', function() {
     var typed2 = typed.create();
-    typed2.conversions = [
+    typed2.addConversion(
       {from: 'Date', to: 'string', convert: function (x) {return x.toISOString()}}
-    ];
+    );
 
     var fn = typed2({
       'string': function () {
@@ -293,7 +306,23 @@ describe('conversion', function () {
     assert.equal(fn(2, 4, 5), 'numbers');
   });
 
-  it('should not apply conversions when having an any type argument', function() {
+  it('should only end up with one way to convert a signature', function() {
+    var t2 = typed.create()
+    t2.addConversions([
+      {from: 'number', to: 'string', convert: n => 'N' + n},
+      {from: 'Array', to: 'boolean', convert: A => A.length > 0}
+    ]);
+    var ambiguous = t2({
+      'string, Array': (s, A) => 'one ' + s,
+      'number, boolean': (n, b) => 'two' + n,
+    }); // Could be two ways to apply to 'number, Array'; want only one
+    assert.strictEqual(ambiguous._typedFunctionData.signatures.length, 3)
+    assert.strictEqual(
+      t2.find(ambiguous, 'number, Array').apply(null, [0, [0]]),
+      'two0');
+  });
+
+  it('should prefer conversions to any type argument', function() {
     var fn = typed({
       'number': function (a) {
         return 'number';
@@ -304,19 +333,35 @@ describe('conversion', function () {
     });
 
     assert.equal(fn(2), 'number');
-    assert.equal(fn(true), 'any');
+    assert.equal(fn(true), 'number');
     assert.equal(fn('foo'), 'any');
     assert.equal(fn('{}'), 'any');
+  });
+
+  it('should allow removal of conversions', function () {
+    const inc = typed({number: n => n + 1});
+    assert.strictEqual(inc(true), 2);
+    typed.removeConversion({
+      from: 'boolean',
+      to: 'number',
+      convert: convertBool
+    });
+    assert.throws(() => typed.convert(false, 'number'), /no conversions/);
+    const dec = typed({number: n => n - 1});
+    assert.throws(() => dec(true), /TypeError: Unexpected type/);
+    // But pre-existing functions remain OK:
+    assert.strictEqual(inc(true), 2);
   });
 
   describe ('ordering', function () {
 
     it('should correctly select the signatures with the least amount of conversions', function () {
-      typed.conversions = [
+      typed.clearConversions();
+      typed.addConversions([
         {from: 'boolean', to: 'number', convert: function (x) {return +x;}},
         {from: 'number',  to: 'string', convert: function (x) {return x + '';}},
         {from: 'boolean', to: 'string', convert: function (x) {return x + '';}}
-      ];
+      ]);
 
       var fn = typed({
         'boolean, boolean': function (a, b) {
@@ -346,18 +391,18 @@ describe('conversion', function () {
       assert.equal(fn(true, 'foo'), 'strings');
       assert.equal(fn('foo', true), 'strings');
 
-      assert.deepEqual(Object.keys(fn.signatures), [
-        'number,number',
-        'string,string',
-        'boolean,boolean'
-      ]);
+      assert.equal(Object.keys(fn.signatures).length, 3);
+      assert.ok('number,number' in fn.signatures);
+      assert.ok('string,string' in fn.signatures);
+      assert.ok('boolean,boolean' in fn.signatures);
     });
 
     it('should select the signatures with the conversion with the lowest index (1)', function () {
-      typed.conversions = [
+      typed.clearConversions();
+      typed.addConversions([
         {from: 'boolean', to: 'string', convert: function (x) {return x + '';}},
         {from: 'boolean', to: 'number', convert: function (x) {return x + 0;}}
-      ];
+      ]);
 
       // in the following typed function, a boolean input can be converted to
       // both a string or a number, which is both ok. In that case,
@@ -370,17 +415,17 @@ describe('conversion', function () {
 
       assert.strictEqual(fn(true), 'true');
 
-      assert.deepEqual(Object.keys(fn.signatures), [
-        'number',
-        'string'
-      ]);
+      assert.equal(Object.keys(fn.signatures).length, 2);
+      assert.ok('number' in fn.signatures);
+      assert.ok('string' in fn.signatures);
     });
 
     it('should select the signatures with the conversion with the lowest index (2)', function () {
-      typed.conversions = [
+      typed.clearConversions();
+      typed.addConversions([
         {from: 'boolean', to: 'number', convert: function (x) {return x + 0;}},
         {from: 'boolean', to: 'string', convert: function (x) {return x + '';}}
-      ];
+      ]);
 
       // in the following typed function, a boolean input can be converted to
       // both a string or a number, which is both ok. In that case,
@@ -395,11 +440,12 @@ describe('conversion', function () {
     });
 
     it('should select the signatures with least needed conversions (1)', function () {
-      typed.conversions = [
+      typed.clearConversions();
+      typed.addConversions([
         {from: 'number', to: 'boolean', convert: function (x) {return !!x }},
         {from: 'number', to: 'string', convert: function (x) {return x + '' }},
         {from: 'boolean', to: 'string', convert: function (x) {return x + '' }}
-      ];
+      ]);
 
       // in the following typed function, the number input can be converted to
       // both a string or a boolean, which is both ok. It should pick the
@@ -413,11 +459,12 @@ describe('conversion', function () {
     });
 
     it('should select the signatures with least needed conversions (2)', function () {
-      typed.conversions = [
+      typed.clearConversions();
+      typed.addConversions([
         {from: 'number', to: 'boolean', convert: function (x) {return !!x }},
         {from: 'number', to: 'string', convert: function (x) {return x + '' }},
         {from: 'boolean', to: 'string', convert: function (x) {return x + '' }}
-      ];
+      ]);
 
       // in the following typed function, the number input can be converted to
       // both a string or a boolean, which is both ok. It should pick the